feat: small enhancement
