+++ PrebidMobile/AdUnits/Native/NativeAd.swift
@@ -147,9 +147,7 @@ public class NativeAd: NSObject, CacheExpiryDelegate {
             return nil
         }
         
-        let macrosHelper = PBMORTBMacrosHelper(bid: rawBid)
-        rawBid.adm = macrosHelper.replaceMacros(in: rawBid.adm)
-        rawBid.nurl = macrosHelper.replaceMacros(in: rawBid.nurl)
+        let bid = Bid(bid: rawBid)
         
         let ad = NativeAd()
         ad.bid = bid
@@ -377,13 +375,12 @@ public class NativeAd: NSObject, CacheExpiryDelegate {
     }
     
     private func fireClickTrackers() {
-        guard let clickTrackersURLs = nativeAdMarkup?.link?.clicktrackers else { return }
-        
-        if clickTrackersURLs.count > 0 {
-            TrackerManager.shared.fireTrackerURLArray(arrayWithURLs: clickTrackersURLs) {
-                _ in
-            }
+        guard let clickTrackersURLs = nativeAdMarkup?.link?.clicktrackers,
+              clickTrackersURLs.count > 0 else {
+            return
         }
+        
+        TrackerManager.shared.fireTrackerURLArray(arrayWithURLs: clickTrackersURLs) { _ in }
     }
     
     private func presentSKStoreProductViewController(with productParameters: [String: Any]) {
